+++ CHANGELOG.md
@@ -5,10 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.3] - 2021-03-14
+
+### Changed
+- Added functionality that will copy Content between languages and have it ready for Translation review.
+
+
 ## [1.1.2] - 2021-02-23
 
 ### Changed
-- Code that will adjust editorial workflow and set that every stage of process is considered a default revision.
+- Code that will adjust the editorial workflow and set that every stage of process is considered a default revision.
 
 
 ## [1.1.1] - 2021-02-15